refactor and fix some tests
