+++ imod_coupler/utils.py
@@ -13,7 +13,7 @@ from numpy.typing import NDArray
 from scipy.sparse import csr_matrix
 
 from imod_coupler.config import LogLevel
-
+from imod_coupler.logging.exchange_collector import ExchangeCollector
 
 def create_mapping(
     src_idx: Any,
@@ -88,6 +88,77 @@ def setup_logger(log_level: LogLevel, log_file: Path) -> None:
     # Add handler for file
     log_file.unlink(missing_ok=True)
     logger.add(log_file, level=log_level)
+
+
+class MemoryExchange:
+    """Class to handle n:m exchanges between two pointers arrays"""
+
+    def __init__(
+        self,
+        ptr_a: NDArray[np.float64 | np.int32],
+        ptr_b: NDArray[np.float64 | np.int32],
+        ptr_a_index: NDArray[np.int32],
+        ptr_b_index: NDArray[np.int32],
+        exchange_logger: ExchangeCollector,
+        ptr_a_conversion: NDArray[np.float64] | None = None,
+        ptr_b_conversion: NDArray[np.float64] | None = None,
+        exchange_operator: str | None = None,
+    ) -> None:
+        self.ptr_a = ptr_a
+        self.ptr_b = ptr_b
+        self.ptr_a_index = ptr_a_index
+        self.ptr_b_index = ptr_b_index
+        self.exchange_logger = exchange_logger
+        self.exchange_operator = exchange_operator
+        self._set_conversion_terms(ptr_a_conversion, ptr_b_conversion)
+        self.mapping, self.mask = create_mapping(
+            self.ptr_a_index,
+            self.ptr_b_index,
+            self.ptr_a.size,
+            self.ptr_b.size,
+            self.exchange_operator,
+            self.conversion_term,
+        )
+
+    def _set_conversion_terms(
+        self,
+        ptr_a_conversion: NDArray[np.float64] | None = None,
+        ptr_b_conversion: NDArray[np.float64] | None = None,
+    ) -> None:
+        if ptr_a_conversion is None and ptr_b_conversion is None:
+            self.conversion_term = None
+        else:
+            conversion_term = np.ones_like(self.ptr_a_index, dtype=np.float64)
+            if ptr_a_conversion is not None:
+                self._raise_if_not_compatible(ptr_a_conversion, self.ptr_a)
+                conversion_term = ptr_a_conversion[self.ptr_a_index]
+            if ptr_b_conversion is not None:
+                self._raise_if_not_compatible(ptr_b_conversion, self.ptr_b)
+                conversion_term = conversion_term * ptr_b_conversion[self.ptr_b_index]
+            self.exchange_operator = (
+                None  # effect of operator should be in conversion term
+            )
+            self.conversion_term = conversion_term
+
+    def _raise_if_not_compatible(self, array1:NDArray[np.float64], array2:NDArray[np.float64]) -> None:
+        if array1.shape != array2.shape:
+            raise ValueError(
+                "conversion array should have the same shape as corresponding ptr"
+            )
+        
+    def exchange(self, delt: np.float64 = 1.0) -> None:
+        """Exchange Kernel a to Kernel b"""
+        self.ptr_b[:] = self.mask[:] * self.ptr_b[:] + self.mapping.dot(self.ptr_a)[:] / delt
+
+    def log(self, label: str, time: float) -> None:
+        """Log the exchange for receiving side; array b"""
+        self.exchange_logger.log_exchange(label, self.ptr_b[:], time)
+        self.label = label
+
+    def finalize_log(self) -> None:
+        """finalizes the exchange within the logger, if present"""
+        if self.label in self.exchange_logger.exchanges.keys():
+            self.exchange_logger.exchanges[self.label].finalize()
 
 
 @contextmanager
